+++ src/app/application/app.component.html
@@ -7,10 +7,11 @@
       <dng-sidemenu></dng-sidemenu>
     </div>
     <div class="col-xs-12 col-sm-10">
-        <ol class="breadcrumb">
+        <!--ol class="breadcrumb">
             <li class="breadcrumb-item"><a routerLink="/home">Home</a></li>
             <li class="breadcrumb-item active">Courses</li>
-        </ol>
+        </ol-->
+        <dng-breadcrumb></dng-breadcrumb>
         <div class="jumbotron">
           <router-outlet></router-outlet>
         </div>
