+++ src/app/widgets/menu/breadcrumb.component.html
@@ -1,0 +1,4 @@
+<ol class="breadcrumb">
+    <li class="breadcrumb-item"><a routerLink="/home">Home</a></li>
+    <li class="breadcrumb-item active">Courses</li>
+</ol>
