+++ src/app/widgets/menu/index.ts
@@ -1,1 +1,2 @@
 export * from './sidemenu.component';
+export * from './breadcrumb.component';
